split app container to add pull request list container
