+++ app/src/main/java/com/sample/pulls/di/PullRequestListContainer.kt
@@ -1,0 +1,10 @@
+package com.sample.pulls.di
+
+import com.sample.pulls.domain.repositories.GithubRepository
+import com.sample.pulls.presentation.pullslist.viewmodel.PullRequestListViewModelFactory
+
+class PullRequestListContainer(githubRepository: GithubRepository) {
+    init {
+        PullRequestListViewModelFactory.inject(githubRepository)
+    }
+}
